+++ AutomaticAnalysis/Configurations/configuration_template.json
@@ -1,31 +1,51 @@
 {
-        "CONFIGURATION_NAME": "", 
-        "RECORD_REF_FIELDNAME": "accessURI",
-        "SKIP_LIST": ["Image Name", "catalogNumber", "Dataset Source", "accessURI", "Label only?", "modifiedBy", "verifiedBy" , "substrate", "URL", "Image"], 
-        "CORE_FIELDS_LIST": [],
-        "USE_CORE_FIELDS": "False",
-        "EDIT_DISTANCE_CLASS": "",
-        "USE_FIELDNAMES_EDIT_DISTANCE": "False",
-        "EDIT_DISTANCE_CONFIG": {"comment1": "COSTS are lists of lists.",
-                                 "comment2": "For INSERT_CHAR and DELETE_CHAR, the order of the two elements of the inner list is char, cost.",
-                                 "comment3": "For SUBSTITUTION_CHAR and TRANSPOSITION_CHAR, the order of the three elements of the inner list is char to substitute in, target char to be substuted out, cost.",
-                                 "INSERT_CHAR_COSTS": [[]],  
-                                 "DELETE_CHAR_COSTS":  [[]],
-                                 "SUBSTITUTION_CHAR_COSTS": [[]],
-                                 "TRANSPOSITON_CHAR_COSTS": [[]]
-                                },
-        "FIELDNAMES_COSTS": {"": {"INSERT_CHAR_COSTS": [[]],  
-                                  "DELETE_CHAR_COSTS":  [[]],
-                                  "SUBSTITUTION_CHAR_COSTS": [[]],
-                                  "TRANSPOSITON_CHAR_COSTS": [[]]
-                                }
-                            },                       
-        "USE_TOLERANCES": "False",
-        "TOLS": {"": []},
-        "LLM_SPREAD_SOURCES": [],
-        "GROUND_TRUTH_FILENAME": "",
-        "RESULT_FILENAME": "",
-        "ERRORS_FILENAME": "",
-        "SOURCE_PATH": "AccuracyTesting/AccuracyTestingSources/",
-        "RESULTS_PATH": "AccuracyTesting/AccuracyTestingResults/"
+  "comment1": "Create an empty .json file in the Configurations directory with a name of your choosing.",
+  "comment2": "Copy the contents of this file into empty file.",
+  "comment3": "Fill out the copied configuration.",
+  "comment4": "See how_to_make_a_configuration.md for instructions and explanations.", 
+  "comment5": "Lastly, enter the configuration filename into comparison_and_accuracy.py and your ready to roll!!!!",
+  "CONFIGURATION_NAME": "",
+
+  "COMPARISON_CONFIG": {"RECORD_REF_FIELDNAME": "accessURI",
+                          "SKIP_LIST": ["Image Name", "catalogNumber", "Dataset Source", "accessURI", "Label only?", "modifiedBy", "verifiedBy" , "substrate", "URL", "Image"], 
+                          "CORE_FIELDS_LIST": [],
+                          "USE_CORE_FIELDS": "",
+                          "LLM_SPREAD_SOURCES": [""],
+                          "GROUND_TRUTH_FILENAME": "",
+                          "RESULT_FILENAME": "",
+                          "ERRORS_FILENAME": "",
+                          "SOURCE_PATH": "AutomaticAnalysis/Sources/",
+                          "RESULTS_PATH": "AutomaticAnalysis/Results/"
+  },
+  
+  "EDIT_DISTANCE_CONFIG": {
+                           "comment1": "COSTS are lists of lists.",
+                           "comment2": "For INSERT_CHAR and DELETE_CHAR, the order of the two elements of the inner list is char, cost.",
+                           "comment3": "For SUBSTITUTION_CHAR and TRANSPOSITION_CHAR, the order of the three elements of the inner list is char to substitute in, target char to be substuted out, cost.",
+                           "ALL_FIELDS_CUSTOM_COSTS": {
+                                              "INSERT_CHAR_COSTS": [[]],  
+                                              "DELETE_CHAR_COSTS":  [[]],
+                                              "SUBSTITUTION_CHAR_COSTS": [[]],
+                                              "TRANSPOSITON_CHAR_COSTS": [[]]
+
+                           },
+                           "USE_FIELDNAMES_EXCLUSIVELY": "",
+                           "FIELDNAMES_COSTS": {
+                                                "": {"INSERT_CHAR_COSTS": [[]],  
+                                                    "DELETE_CHAR_COSTS":  [[]],
+                                                    "SUBSTITUTION_CHAR_COSTS": [[]],
+                                                    "TRANSPOSITON_CHAR_COSTS": [[]]
+                                                    }
     }
+                          },
+
+  "TOLERANCES_CONFIG": {"TOLERANCES_ALLOWED": "",
+                        "USE_GRADED_MATCH_THRESHOLD": "",
+                        "GRADED_MATCH_THRESHOLD": "0.99999",
+                        "TOLS": {"": [""]
+                                 }
+
+  }
+  
+  
+}
