--- AutomaticAnalysis/Results/demo.csv
@@ -1,2 +1,0 @@
-run,ground truth filename,configuration name,errors filename,verbatimCollectors,collectedBy,secondaryCollectors,recordNumber,verbatimEventDate,minimumEventDate,maximumEventDate,verbatimIdentification,latestScientificName,identifiedBy,verbatimDateIdentified,associatedTaxa,country,firstPoliticalUnit,secondPoliticalUnit,municipality,verbatimLocality,locality,habitat,verbatimElevation,verbatimCoordinates,otherCatalogNumbers,originalMethod,typeStatus,TrueApplicable,TrueN/A,FalseApplicable,FalseN/A,GradedTA,GradedFA,Correct:TA+TN/A,Errors:FA+FN/A,TA+TNA/TA+TNA+FA+FNA,acc1:TA+TNA/TA+TNA+FA+FNA,grdTA+TNA/grdTA+TNA+grdFA+FNA,acc2:grdTP+TN/grdTP+TN+grdFP+FN,TA/TA+FA+FNA,acc3:TA/TA+FA+FNA,grdTA/grdTA+grdFA+FNA,acc4:grdTA/grdTA+grdFA+FNA
-TolsSpread_6_11_1050.csv,First100BryophytesTyped.csv,demonstration_only,demo_errors.txt,"[45, 45, 100]","[69, 69, 100]","[83, 83, 100]","[72, 72, 100]","[71, 71, 100]","[80, 80, 100]","[100, 100, 100]","[64, 64, 100]","[77, 77, 100]","[53, 53, 84]","[74, 74, 89]","[94, 94, 100]","[99, 99, 100]","[83, 83, 100]","[85, 85, 100]","[75, 75, 100]","[8, -1129.9, 100]","[100, 100, 100]","[27, 27, 100]","[47, 47, 100]","[68, 68, 100]","[82, 82, 100]","[96, 96, 100]","[100, 100, 100]",1201,551,553,68,63.11708432246762,1690.8829156775323,1752,621,1201+551/1201+551+553+68,0.7383059418457648,63+551/63+551+1690+68,0.2588532883642496,1201/1201+553+68,0.6591657519209659,63/63+1690+68,0.03459637561779242
