--- AutomaticAnalysis/Results/tolerances_experiment.csv
@@ -1,2 +1,0 @@
-run,ground truth filename,configuration name,errors filename,verbatimCollectors,collectedBy,secondaryCollectors,recordNumber,verbatimEventDate,minimumEventDate,maximumEventDate,verbatimIdentification,latestScientificName,identifiedBy,verbatimDateIdentified,associatedTaxa,country,firstPoliticalUnit,secondPoliticalUnit,municipality,verbatimLocality,locality,habitat,verbatimElevation,verbatimCoordinates,otherCatalogNumbers,originalMethod,typeStatus,TrueApplicable,TrueN/A,FalseApplicable,FalseN/A,GradedTA,GradedFA,Correct:TA+TN/A,Errors:FA+FN/A,TA+TNA/TA+TNA+FA+FNA,acc1:TA+TNA/TA+TNA+FA+FNA,grdTA+TNA/grdTA+TNA+grdFA+FNA,acc2:grdTP+TN/grdTP+TN+grdFP+FN,TA/TA+FA+FNA,acc3:TA/TA+FA+FNA,grdTA/grdTA+grdFA+FNA,acc4:grdTA/grdTA+grdFA+FNA
-TolsSpread_6_11_1050.csv,First100BryophytesTyped.csv,tolerances_experiment,tolerances_experiment_errors.txt,"[45, 45, 100]","[69, -867.1, 100]","[83, 83, 100]","[72, 72, 100]","[71, 71, 100]","[80, 80, 100]","[100, 100, 100]","[64, 64, 100]","[77, 77, 100]","[53, 53, 84]","[74, 74, 89]","[94, 94, 100]","[99, 99, 100]","[83, 83, 100]","[85, 85, 100]","[75, 75, 100]","[8, 8, 100]","[9, 9, 100]","[27, 27, 100]","[47, 47, 100]","[68, 68, 100]","[82, 82, 100]","[96, 96, 100]","[100, 100, 100]",1110,551,644,68,173.86695208386394,1580.1330479161354,1661,712,1110+551/1110+551+644+68,0.6999578592498946,173+551/173+551+1580+68,0.30522765598650925,1110/1110+644+68,0.6092206366630076,173/173+1580+68,0.09500274574409664
