+++ AutomaticAnalysis/run_variability.py
@@ -11,8 +11,8 @@
 
 import utility
 
-SOURCE_PATH = "AutomaticAnalysis/SourcesForPaper/"
-RESULTS_PATH = "AutomaticAnalysis/ResultsForPaper/"
+SOURCE_PATH = "AutomaticAnalysis/Sources/"
+RESULTS_PATH = "AutomaticAnalysis/Results/"
 
 # FILENAMES are iterated through as a list of lists
 # The variability will be calculated for each possible pairing of files of each inner list.
